modify by Xingyu

+++ src/com/system/model/order/OrderLine.java
@@ -1,4 +1,5 @@
 package com.system.model.order;
+import java.math.BigDecimal;
 import java.util.*;
 import com.system.model.product.abstractproduct.*;
 
@@ -7,7 +8,7 @@ public class OrderLine {
 	Product orderedAs;
 	private String productCode;
 	private Integer quantity;
-	private Float lineCost;
+	private BigDecimal lineCost;
 
 	public void calculateLineCost() {
 		// TODO - implement OrderLine.calculateLineCost
